ui: rename folder
